+++ src/main/java/com/muhammaddaffa/nextcredits/events/CreditReceiveEvent.java
@@ -4,7 +4,7 @@ import org.bukkit.entity.Player;
 
 public class CreditReceiveEvent extends CreditEvent {
 
-    public CreditReceiveEvent(Player who, int amount) {
+    public CreditReceiveEvent(Player who, double amount) {
         super(who, amount);
     }
